+++ timeago_test.go
@@ -121,3 +121,31 @@ func TestParseFuncWillCalculateIntervalToFutureDate(t *testing.T) {
 		})
 	}
 }
+
+func TestOptionIsEnabled(t *testing.T) {
+	t.Run("returns true if option is enabled", func(test *testing.T) {
+		globalOptions = []string{"noSuffix"}
+
+		if res := optionIsEnabled("noSuffix"); res == false {
+			test.Error("Result must be true, but got false instead")
+		}
+
+		globalOptions = []string{}
+	})
+
+	t.Run("returns true if option is enabled with other option", func(test *testing.T) {
+		globalOptions = []string{"noSuffix", "upcoming"}
+
+		if res := optionIsEnabled("upcoming"); res == false {
+			test.Error("Result must be true, but got false instead")
+		}
+
+		globalOptions = []string{}
+	})
+
+	t.Run("returns false if option is disabled", func(test *testing.T) {
+		if res := optionIsEnabled("noSuffix"); res == true {
+			test.Error("Result must be true, but got false instead")
+		}
+	})
+}