+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ----
 
+## v2.1.7 (2023-09-08)
+
+- Renamed master branch to main
+- Added more tests
+- Simplified code
+- Replaced deprecated `ioutil.ReadFile` with `os.ReadFile`
+
+----
+
 ## v2.1.6 (2022-11-27)
 
 - Changed typo in README.md file;